Train classifier

+++ scripts/generate_data.py
@@ -35,9 +35,10 @@ def generate_data():
     np.savez(
         f"databases/semigroup.{cardinality}.npz",
         database=np.stack(list(positive.values()) + list(negative.values())),
-        labels=np.concatenate(
-            [np.ones(len(positive)), np.zeros(len(negative))]
-        )
+        labels=np.concatenate([
+            np.ones(len(positive), dtype=np.int64),
+            np.zeros(len(negative), dtype=np.int64)
+        ])
     )
